smart-study-buddy/smart-study-buddy/requirements.txt
# Core dependencies
openai>=1.12.0
anthropic>=0.18.0
python-dotenv>=1.0.0

# Vector database (optional for future features)
pinecone-client>=3.0.0

# Web framework (for API/web interface)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0

# CLI and UI
rich>=13.7.0
typer>=0.9.0
gradio>=4.16.0

# Utilities
requests>=2.31.0
pytest>=8.0.0
black>=24.1.0
